TMPluginController -version appears to have changed to a CGFloat. Not sure if this will cause problems in TM1 :\ This is the first step to addressing #8.

+++ source/TMPlugInController.h
@@ -10,6 +10,6 @@
 
 @protocol TMPlugInController <NSObject>
 
-- (float)version;
+- (CGFloat)version;
 
 @end
